(wip) move sensor map to separate file
